+++ src/main/java/com/devbliss/gpullr/controller/UserController.java
@@ -1,28 +1,29 @@
 package com.devbliss.gpullr.controller;
 
-
-import com.devbliss.gpullr.service.github.GithubService;
-import java.io.IOException;
+import com.devbliss.gpullr.domain.User;
+import com.devbliss.gpullr.service.UserService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+/**
+ * REST controller to manage users.
+ * 
+ * @author Henning Schütz <henning.schuetz@devbliss.com>
+ *
+ */
 @RestController
-@RequestMapping("/sample")
-public class SampleController {
-  
+@RequestMapping("/users")
+public class UserController {
+
   @Autowired
-  private GithubService githubService;
+  private UserService userService;
   
   @RequestMapping(method=RequestMethod.GET)
-  public List<String> getShasOfDementity() {
-    try {
-      return githubService.proofThatItWorks();
-    } catch (IOException e) {
-      e.printStackTrace();
-      return null;
-    }
+  public List<User> getAllUsers() {
+    return userService.findAll();
   }
+
 }