+++ src/main/java/com/devbliss/gpullr/repository/PersistenceConfig.java
@@ -41,7 +41,7 @@ public class PersistenceConfig {
   public JpaVendorAdapter createJpaVendorAdapter() {
     HibernateJpaVendorAdapter adapter = new HibernateJpaVendorAdapter();
     adapter.setDatabase(Database.H2);
-    adapter.setShowSql(true);
+    adapter.setShowSql(false);
     adapter.setGenerateDdl(true);
     adapter.setDatabasePlatform("org.hibernate.dialect.H2Dialect");
     return adapter;