+++ src/main/java/com/devbliss/gpullr/service/UserService.java
@@ -1,0 +1,32 @@
+package com.devbliss.gpullr.service;
+
+import com.devbliss.gpullr.domain.User;
+import com.devbliss.gpullr.repository.UserRepository;
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+/**
+ * Business Layer for {@link com.devbliss.gpullr.domain.User} objects.
+ * 
+ * @author Henning Schütz <henning.schuetz@devbliss.com>
+ *
+ */
+@Service
+public class UserService {
+
+  private final UserRepository userRepository;
+  
+  @Autowired
+  public UserService(UserRepository userRepository) {
+    this.userRepository = userRepository;
+  }
+
+  public void insertOrUpdate(User user) {
+    userRepository.save(user);
+  }
+
+  public List<User> findAll() {
+    return userRepository.findAll();
+  }
+}