+++ changelog.md
@@ -1,5 +1,6 @@
 
 ### Version 0.1.0-SNAPSHOT
+ - [feature] periodically fetching users, repos and (PR-)events
  - [bug] fixed database config
  - [feature] github service / jcabi access setup
  - [feature] better test output for gradle
@@ -8,4 +9,4 @@
  - [feature] project setup
  - [initial] initial commit
 
--- Last change by: hschuetz <henning.schuetz@devbliss.com> Wed Feb 11 10:28:35 CET 2015
+-- Last change by: hschuetz <henning.schuetz@devbliss.com> Mon Feb 16 11:06:38 CET 2015